Fixes for a range of issue reported by the Intel C Compiler. IIRC, this covers warnings over converting 64 bit values to lower precision variables, hiding variable names, and an unused variable.
svn path=/trunk/kdesupport/qca/; revision=535284

+++ src/support/dirwatch/dirwatch.cpp
@@ -30,7 +30,7 @@ static int platform_ref = 0;
 struct file_info
 {
 	QString name;
-	uint size;
+	qint64 size;
 	QDateTime lastModified;
 };
